fix(lenses): third Cockroach dialect error, fail-safe 503s, job heartbeats, phantom-job schedules
Four audit P0s:

- chain_coverage's COALESCE(float/int) threw on CockroachDB (code 22023,
  live in prod logs) — the third dialect error hiding behind the jsonb
  fixes in #13. Cast both operands ::float8 with a 0.0 fallback, and fix
  the same pattern in baselines, stable-corridors, and trend.

- Degraded lenses now return HTTP 503 (lens_degraded) instead of
  200 + zeros + degraded flag: clients that predate the degraded
  contract render zeros as "Covered — OK", so fail-safe cannot depend
  on a client upgrade. Current MCP renders the 503 as the standard
  degraded notice via renderIfDegraded503.

- job_heartbeats (migration 000025): every cron run upserts a heartbeat;
  freshness probes read the heartbeat for pipeline health and keep data
  recency as a separate neutral signal. Kills the false-red where a
  quiet network reported "aggregation stale" minutes after a green run.

- Schedule the phantom jobs: data-archival + agent-expiration (daily),
  watch-evaluation + pattern-detection (2x/hour). Their handlers existed
  but no scheduler called them — set_watch never fired and the privacy
  policy's retention promises never executed.

+++ packages/ingestion-api/src/routes/baselines.ts
@@ -104,12 +104,12 @@ app.get('/baselines/cohort', async (c) => {
               PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY duration_ms)::float AS "median_duration_ms",
               PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY duration_ms)::float AS "p95_duration_ms",
               COALESCE(
-                COUNT(*) FILTER (WHERE status = 'failure')::float /
-                NULLIF(COUNT(*), 0), 0
+                COUNT(*) FILTER (WHERE status = 'failure')::float8 /
+                NULLIF(COUNT(*), 0)::float8, 0.0
               ) AS "failure_rate",
               COALESCE(
-                COUNT(*) FILTER (WHERE anomaly_flagged = true)::float /
-                NULLIF(COUNT(*), 0), 0
+                COUNT(*) FILTER (WHERE anomaly_flagged = true)::float8 /
+                NULLIF(COUNT(*), 0)::float8, 0.0
               ) AS "anomaly_rate"
        FROM interaction_receipts
        WHERE emitter_provider_class = $1
